feat(protocol): Move MetricUnit to common protocol module
